Fix export type
