+++ lib/rsa-xml.js
@@ -127,6 +127,12 @@ RSAXML.prototype._parseXMLKey = function (key) {
  * Module exports
  */
 
-module.exports = function(key, container) {
+ const buildRSAXML = function (key, container) {
     return new RSAXML(key, container);
 };
+
+buildRSAXML.default = function (key, container) {
+    return new RSAXML(key, container);
+};
+
+module.exports = buildRSAXML;
