+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.3",
   "description": "RSA XML Decrypter",
   "main": "index.js",
+  "types": "index.d.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
